# Refactor and optimize yuv420p and nv12 processing