… test/torchaudio_unittest/io/stream_reader_test.py